webapp/requirements.txt
streamlit
requests
pillow
opencv-python
numpy